backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docflow-studio"
version = "0.1.0"
description = "Enterprise-grade document ingestion and preprocessing for RAG pipelines."
readme = "README.md"
requires-python = ">=3.13"
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "aiosqlite>=0.20.0,<1.0.0",
  "pydantic>=2.9.0,<3.0.0",
  "pydantic-settings>=2.5.0,<3.0.0",
  "python-multipart>=0.0.9,<1.0.0",
  "orjson>=3.10.0,<4.0.0",
  "pypdf>=5.0.0,<6.0.0",
  "python-docx>=1.1.0,<2.0.0",
  "beautifulsoup4>=4.12.0,<5.0.0",
  "ftfy>=6.3.0,<7.0.0",
  "langdetect>=1.0.9,<2.0.0",
  "tenacity>=9.0.0,<10.0.0",
  "structlog>=24.4.0,<25.0.0",
  "httpx>=0.27.0,<1.0.0",
  "mcp>=1.9.0,<2.0.0",
  "pyyaml>=6.0.2,<7.0.0",
  "pywebview>=6.2.1,<7.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
  "pytest-cov>=5.0.0,<6.0.0",
  "ruff>=0.7.0,<1.0.0",
  "pyinstaller>=6.10.0,<7.0.0",
]

[project.scripts]
docflow-api = "docflow.cli:run_api"
docflow-worker = "docflow.cli:run_worker"
docflow-mcp = "docflow.cli:run_mcp"
docflow-desktop = "docflow.desktop_launcher:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
